% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeRect}
\alias{makeRect}
\title{makeRect auxiliary function}
\usage{
makeRect(x, y, w, h, ...)
}
\arguments{
\item{x}{center x coord.}

\item{y}{center y coord.}

\item{w}{width.}

\item{h}{height.}

\item{...}{optional arguments passed to \code{rect()}.}
}
\value{
NULL function called for its plotting side effects
}
\description{
draws a rectangle centered on the given coordinates and of specified width and height
}
\details{

}

