% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ThanoProjMatrix}
\alias{ThanoProjMatrix}
\title{ThanoProjMatrix a super cheap Lesile matrix for thanatological model}
\usage{
ThanoProjMatrix(fy, da, lambda)
}
\arguments{
\item{fy}{fertility rates classified by thanatological age.}

\item{da}{lifetable deaths distribution.}

\item{lambda}{discount fertility the partof infant mortality that doesn't make it to the end of the year.}
}
\value{
the projection matrix
}
\description{
This could be much more thoughtful.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

